refactoring: 유저 인가 기능 추가

+++ matzipback/src/main/java/com/matzip/matzipback/board/command/application/controller/PostCmtCommandController.java
@@ -3,6 +3,9 @@ package com.matzip.matzipback.board.command.application.controller;
 import com.matzip.matzipback.board.command.application.dto.ReqPostCmtCreateDTO;
 import com.matzip.matzipback.board.command.application.dto.ReqPostCmtUpdateDTO;
 import com.matzip.matzipback.board.command.application.service.PostCommentService;
+import com.matzip.matzipback.common.util.CustomUserUtils;
+import com.matzip.matzipback.exception.RestApiException;
+import com.matzip.matzipback.responsemessage.SuccessCode;
 import com.matzip.matzipback.responsemessage.SuccessResMessage;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -11,6 +14,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import static com.matzip.matzipback.exception.ErrorCode.FORBIDDEN_ACCESS;
+import static com.matzip.matzipback.exception.ErrorCode.UNAUTHORIZED_REQUEST;
 import static com.matzip.matzipback.responsemessage.SuccessCode.*;
 
 @RestController
@@ -25,8 +30,16 @@ public class PostCmtCommandController {
     @PostMapping("/postcomment")
     @Operation(summary = "게시글 댓글 등록", description = "게시글에 댓글을 등록한다.")
     public ResponseEntity<SuccessResMessage> createPostComment(@Valid @RequestBody ReqPostCmtCreateDTO reqPostCmtCreateDTO) {
-        postCommentService.createPostComment(reqPostCmtCreateDTO);
-        return ResponseEntity.ok(new SuccessResMessage(BASIC_SAVE_SUCCESS));
+
+        try{
+            if (CustomUserUtils.getCurrentUserAuthorities().iterator().next().getAuthority().equals("user")) {
+
+                postCommentService.createPostComment(reqPostCmtCreateDTO);
+                return ResponseEntity.ok(new SuccessResMessage(BASIC_SAVE_SUCCESS));
+            } else {
+                throw new RestApiException(FORBIDDEN_ACCESS);
+            }
+        } catch (NullPointerException e) { throw new RestApiException(UNAUTHORIZED_REQUEST); }
     }
 
     // 댓글 수정
@@ -35,16 +48,32 @@ public class PostCmtCommandController {
     public ResponseEntity<SuccessResMessage> updatePostComment(
             @Valid @RequestBody ReqPostCmtUpdateDTO reqPostCmtUpdateDTO
     ) {
-        postCommentService.updatePostComment(reqPostCmtUpdateDTO);
-        return ResponseEntity.ok(new SuccessResMessage(BASIC_UPDATE_SUCCESS));
+
+        try{
+            if (CustomUserUtils.getCurrentUserAuthorities().iterator().next().getAuthority().equals("user")) {
+
+                postCommentService.updatePostComment(reqPostCmtUpdateDTO);
+                return ResponseEntity.ok(new SuccessResMessage(BASIC_UPDATE_SUCCESS));
+            } else {
+                throw new RestApiException(FORBIDDEN_ACCESS);
+            }
+        } catch (NullPointerException e) { throw new RestApiException(UNAUTHORIZED_REQUEST); }
     }
 
     // 댓글 삭제
     @DeleteMapping("/postcomment/{postCommentSeq}")
     @Operation(summary = "게시글 댓글 삭제", description = "게시글에 댓글을 삭제한다.")
     public ResponseEntity<SuccessResMessage> deletePostComment(@PathVariable Long postCommentSeq) {
-        postCommentService.deletePostComment(postCommentSeq);
-        return ResponseEntity.ok(new SuccessResMessage(BASIC_DELETE_SUCCESS));
+
+        try{
+            if (CustomUserUtils.getCurrentUserAuthorities().iterator().next().getAuthority().equals("user")) {
+
+                postCommentService.deletePostComment(postCommentSeq);
+                return ResponseEntity.ok(new SuccessResMessage(BASIC_DELETE_SUCCESS));
+            } else {
+                throw new RestApiException(FORBIDDEN_ACCESS);
+            }
+        } catch (NullPointerException e) { throw new RestApiException(UNAUTHORIZED_REQUEST); }
     }
 
 }
